+++ express-routes/routes.js
@@ -1,8 +1,11 @@
 const { Router } = require("express");
 const axios = require("axios");
 const router = Router();
+const tar = require("tar-stream");
 const apiController = require("../express-controllers/controller");
+const { craigApi } = require("../express-controllers/craig-api");
 const controller = new apiController(axios);
+const craigRoutes = new craigApi(controller, tar);
 
 // vsi
 router.get("/vsi/:region/instanceProfiles", controller.vsiInstanceProfiles);
@@ -18,5 +21,8 @@ router.post(
   "/schematics/:workspaceName/:region?/:resourceGroup?",
   controller.createWorkspace
 );
+
+// craig
+router.post("/craig/tar", craigRoutes.craigTar);
 
 module.exports = router;